mv template def to hpp
